src/requirements.txt
gspread
gspread-formatting
google-auth
beautifulsoup4
lxml
python-dotenv
playwright
google-api-python-client
